Attempt to fix filter
Victoria, Dominique and Shonaigh problem solved and realized the if statements within messageArrived was not working. It was removed and the creation of a booking request and booking registry seem to work but the creation of the dentist registry does not work. The method makeDentistArray is able to take in the MQTT message and create a JSON object from it, but it is unable to access and fill inthe appropriate fields.

Resolves: #3