Add circuit breaker and refactor Filter
Fay and I implemented a working circuit breaker using circuit-breaker-example as a template. We created a new class for handling publish-subscribe which meant that we had to refactor the Filter class quite a bit. We added some comments. It has been tested and works but we need exceptions and to change the settings of the circuit breaker.

Resolves #17